add simple shooting mechanism and destroy enemy when health <= 0

+++ src/components/Animation.hpp
@@ -5,8 +5,8 @@
 #include "utils/Vec.hpp"
 
 struct Animation {
-    uint32_t startTime; // Current animation time since it started
-    uint32_t duration; // In ms
+    uint64_t startTime; // Current animation time since it started
+    uint64_t duration; // In ms
     uint16_t curStep; // Current step of the animation
     uint16_t steps; // Number of steps of the animation
     Vec2i startPos;
