+++ src/systems/AnimationSystem.cpp
@@ -3,7 +3,7 @@
 #include "components/Sprite.hpp"
 #include <iostream>
 
-void animateSprites(entt::registry &reg, uint32_t time)
+void animateSprites(entt::registry &reg, uint64_t time)
 {
     auto view = reg.view<Animation, Sprite>();
 
@@ -16,7 +16,7 @@ void animateSprites(entt::registry &reg, uint32_t time)
             anim.startTime = time;
         }
 
-        const uint32_t timePerFrame = anim.duration / static_cast<uint32_t>(anim.steps);
+        const uint64_t timePerFrame = anim.duration / static_cast<uint64_t>(anim.steps);
         const int currentStep = std::floor((time - anim.startTime) / timePerFrame);
         sprite.textureRect.x = anim.startPos.x + anim.spriteSize.x * currentStep;
         sprite.textureRect.y = anim.startPos.y;
